+++ input_options.go
@@ -6,3 +6,11 @@ type InputOptions struct {
 	Default   string
 	Validator Validator
 }
+
+// SelectOptions are used on the select prompts and allow
+// setting a default index option which is non-zero
+// indexed.
+type SelectOptions struct {
+	Default   int
+	Validator Validator
+}