+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ### Unreleased
 
+## Fixed
+- Fixed an error where select would not account for zero-based indexes during selection. ([#2](https://github.com/pixelandtonic/prompt/issues/2))
+
 ### 1.0.0 - 2020-05-05
 - Initial release.